refactor.header 컴포넌트 분리 및 함수 분리
